feat: finish login refactor first step（some files to be deleted）
