+++ packages/rn-sdk/js/pages/Guardian/GuardianApproval/index.tsx
@@ -19,12 +19,12 @@ import { GuardiansStatus, GuardiansStatusItem } from '../types';
 import { GuardianVerifyConfig } from 'model/verify/social-recovery';
 import { GuardianConfig } from 'model/verify/guardian';
 import { UserGuardianItem } from '@portkey-wallet/store/store-ca/guardians/type';
-import { GuardianApprovalPageResult } from 'components/entries/GuardianApproval';
+import { GuardianApprovalPageResult } from 'pages/entries/GuardianApproval';
 import Loading from 'components/Loading';
 import { verifyHumanMachine } from 'components/VerifyHumanMachine';
 import { guardianTypeStrToEnum, isReacptchaOpen } from 'model/global';
 import { NetworkController } from 'network/controller';
-import { VerifierDetailsPageProps } from 'components/entries/VerifierDetails';
+import { VerifierDetailsPageProps } from 'pages/entries/VerifierDetails';
 import { PortkeyEntries } from 'config/entries';
 import { AccountOriginalType, AfterVerifiedConfig, VerifiedGuardianDoc } from 'model/verify/after-verify';
 import { VerifyPageResult } from '../VerifierDetails';
