+++ packages/rn-sdk/js/pages/Login/LoginPortkey/index.tsx
@@ -10,35 +10,46 @@ import { BGStyles, FontStyles } from 'assets/theme/styles';
 import styles from '../styles';
 import Email from '../components/Email';
 import Phone from '../components/Phone';
+import QRCode from '../components/QRCode';
+import Referral from '../components/Referral';
+import SwitchNetwork from '../components/SwitchNetwork';
 import { PageLoginType } from '../types';
 import GStyles from 'assets/theme/GStyles';
 import fonts from 'assets/theme/fonts';
 import { defaultColors } from 'assets/theme';
-import { CountryCodeItem } from 'types/wallet';
+import useEffectOnce from 'hooks/useEffectOnce';
+import { checkForCountryCodeCached } from 'model/global';
 import useBaseContainer from 'model/container/UseBaseContainer';
 import { PortkeyEntries } from 'config/entries';
+import NetworkContext from '../context/NetworkContext';
+import { NetworkItem } from '@portkey-wallet/types/types-ca/network';
+import { PortkeyConfig, setEndPointUrl } from 'global/constants';
+import { NetworkList } from '@portkey-wallet/constants/constants-ca/network-mainnet';
+import { CountryCodeItem } from 'types/wallet';
+
 const scrollViewProps = { extraHeight: 120 };
 const safeAreaColor: SafeAreaColorMapKeyUnit[] = ['transparent', 'transparent'];
 
-export default function LoginPortkey({
+export default function LoginPortKey({
   selectedCountryCode,
   updateCountryCode,
 }: {
   selectedCountryCode: CountryCodeItem | null;
   updateCountryCode: (item: CountryCodeItem) => void;
 }) {
+  const [loginType, setLoginType] = useState<PageLoginType>(PageLoginType.referral);
+  const [currentNetwork, setCurrentNetwork] = useState<NetworkItem | undefined>(undefined);
+  const { t } = useLanguage();
+  const isMainnet = useMemo(() => {
+    return currentNetwork?.networkType === 'MAIN';
+  }, [currentNetwork?.networkType]);
   const { onFinish } = useBaseContainer({
     entryName: PortkeyEntries.SIGN_IN_ENTRY,
   });
-
-  const [loginType, setLoginType] = useState<PageLoginType>(PageLoginType.phone);
-  const { t } = useLanguage();
-  // const isMainnet = useIsMainnet();
-  const isMainnet = true;
   const loginMap = useMemo(
     () => ({
       [PageLoginType.email]: <Email setLoginType={setLoginType} />,
-      [PageLoginType.qrCode]: <View />,
+      [PageLoginType.qrCode]: <QRCode setLoginType={setLoginType} />,
       [PageLoginType.phone]: (
         <Phone
           setLoginType={setLoginType}
@@ -46,9 +57,31 @@ export default function LoginPortkey({
           updateCountryCode={updateCountryCode}
         />
       ),
+      [PageLoginType.referral]: <Referral setLoginType={setLoginType} />,
     }),
     [selectedCountryCode, updateCountryCode],
   );
+  useEffectOnce(() => {
+    checkForCountryCodeCached();
+    loadCurrentNetwork();
+  });
+  const loadCurrentNetwork = () => {
+    PortkeyConfig.endPointUrl().then(url => {
+      const network = NetworkList.find(item => item.apiUrl === url) || NetworkList[0];
+      setCurrentNetwork(network);
+    });
+  };
+
+  const onBack = () => {
+    if (loginType !== PageLoginType.referral) {
+      setLoginType(PageLoginType.referral);
+    } else {
+      onFinish({
+        status: 'cancel',
+        data: {},
+      });
+    }
+  };
 
   const backgroundImage = useMemo(() => {
     if (isIOS) {
@@ -58,36 +91,42 @@ export default function LoginPortkey({
     }
   }, []);
 
-  const onBack = () => {
-    onFinish({
-      status: 'cancel',
-      data: {},
-    });
+  const networkContext = {
+    currentNetwork: currentNetwork,
+    changeCurrentNetwork: (network: NetworkItem) => {
+      if (network.apiUrl) {
+        setCurrentNetwork(network);
+        setEndPointUrl(network.apiUrl);
+      }
+    },
   };
 
   return (
-    <ImageBackground style={styles.backgroundContainer} resizeMode="cover" source={backgroundImage}>
-      <PageContainer
-        titleDom
-        type="leftBack"
-        themeType="blue"
-        style={BGStyles.transparent}
-        pageSafeBottomPadding={!isIOS}
-        containerStyles={styles.containerStyles}
-        safeAreaColor={safeAreaColor}
-        scrollViewProps={scrollViewProps}
-        leftCallback={onBack}>
-        <Svg icon="logo-icon" size={pTd(60)} iconStyle={styles.logoIconStyle} color={defaultColors.bg1} />
-        <View style={GStyles.center}>
-          {!isMainnet && (
-            <View style={styles.labelBox}>
-              <TextM style={[FontStyles.font11, fonts.mediumFont]}>TEST</TextM>
-            </View>
-          )}
-          <TextXXXL style={[styles.titleStyle, FontStyles.font11]}>{t('Log In To Portkey')}</TextXXXL>
-        </View>
-        {loginMap[loginType as keyof typeof loginMap]}
-      </PageContainer>
-    </ImageBackground>
+    <NetworkContext.Provider value={networkContext}>
+      <ImageBackground style={styles.backgroundContainer} resizeMode="cover" source={backgroundImage}>
+        <PageContainer
+          titleDom
+          type="leftBack"
+          themeType="blue"
+          style={BGStyles.transparent}
+          pageSafeBottomPadding={!isIOS}
+          containerStyles={styles.containerStyles}
+          safeAreaColor={safeAreaColor}
+          scrollViewProps={scrollViewProps}
+          leftCallback={onBack}>
+          <Svg icon="logo-icon" size={pTd(60)} iconStyle={styles.logoIconStyle} color={defaultColors.bg1} />
+          <View style={GStyles.center}>
+            {!isMainnet && (
+              <View style={styles.labelBox}>
+                <TextM style={[FontStyles.font11, fonts.mediumFont]}>TEST</TextM>
+              </View>
+            )}
+            <TextXXXL style={[styles.titleStyle, FontStyles.font11]}>{t('Log In To Portkey')}</TextXXXL>
+          </View>
+          {loginMap[loginType]}
+          <SwitchNetwork />
+        </PageContainer>
+      </ImageBackground>
+    </NetworkContext.Provider>
   );
 }
