+++ packages/rn-sdk/js/pages/entries/GuardianApproval/index.tsx
