+++ packages/rn-sdk/js/pages/entries/SignUp/index.tsx
