+++ packages/rn-sdk/js/pages/entries/VerifierDetails/index.tsx
